backend/requirements.txt
flask
flask_cors
torch
numpy
Pillow
safetensors
transformers
pyngrok
torchvision
matplotlib
scikit-learn
seaborn
tqdm